refactor: extract git functions to src/workflows/develop-git.js
Moves discardWorktreeChanges, glabMrListArgs, fetchOpenPrBranches,
ensureCleanLoopStart, and resetForNextIssue from develop.workflow.js
into a dedicated module. Re-exports ensureCleanLoopStart,
fetchOpenPrBranches, glabMrListArgs, and resetForNextIssue from
develop.workflow.js to preserve existing test import paths.
