+++ src/workflows/develop.workflow.js
@@ -3,7 +3,6 @@ import { randomUUID } from "node:crypto";
 import {
   existsSync,
   mkdirSync,
-  readdirSync,
   readFileSync,
   rmSync,
   writeFileSync,
@@ -17,7 +16,6 @@ import {
 import {
   checkDefaultBranchTracking,
   detectDefaultBranch,
-  detectRemoteType,
   getDefaultBranchRemoteName,
   isStaleUpstreamRefError,
 } from "../helpers.js";
@@ -37,11 +35,8 @@ import qualityReviewMachine from "../machines/develop/quality-review.machine.js"
 import { runPreflight } from "../preflight.js";
 import {
   backupKeyFor,
-  artifactConsistent,
-  clearStateAndArtifacts,
-  saveBackup,
-  restoreBackup,
   prepareForIssue,
+  saveBackup,
 } from "../state/issue-backup.js";
 import { checkpointPathFor } from "../state/machine-state.js";
 import {
@@ -53,6 +48,12 @@ import {
 } from "../state/workflow-state.js";
 import { buildIssueBranchName } from "../worktrees.js";
 import { runHooks, WorkflowRunner } from "./_base.js";
+import {
+  ensureCleanLoopStart,
+  fetchOpenPrBranches,
+  glabMrListArgs,
+  resetForNextIssue,
+} from "./develop-git.js";
 
 /**
  * Update the loop state heartbeat timestamp.
@@ -647,386 +648,14 @@ const isInfraError = (text) =>
     String(text || ""),
   );
 
-/** Unstage, restore tracked files, and remove untracked files. Returns true only if all steps succeeded. */
-function discardWorktreeChanges(repoRoot) {
-  const resetRes = spawnSync("git", ["reset"], {
-    cwd: repoRoot,
-    encoding: "utf8",
-  });
-  if (resetRes.status !== 0) return false;
-
-  const diffRes = spawnSync("git", ["diff", "--name-only"], {
-    cwd: repoRoot,
-    encoding: "utf8",
-  });
-  if (diffRes.status !== 0) return false;
-  const hasTrackedChanges = !!(diffRes.stdout || "").trim();
-  if (hasTrackedChanges) {
-    const coRes = spawnSync("git", ["checkout", "--", "."], {
-      cwd: repoRoot,
-      encoding: "utf8",
-    });
-    if (coRes.status !== 0) return false;
-  }
-
-  const cleanRes = spawnSync("git", ["clean", "-fd", "--exclude=.coder/"], {
-    cwd: repoRoot,
-    encoding: "utf8",
-  });
-  return cleanRes.status === 0;
-}
-
-
-
-
-
-
 // Re-export for tests that import from develop.workflow.js
 export { backupKeyFor, prepareForIssue };
-
-/**
- * Ensure the workspace is in a known-clean state before starting the loop.
- * Cleans up stale per-issue state and artifacts that a previous crashed or
- * interrupted run may have left behind. Does NOT touch loop-state.json so
- * issue-level resume information is preserved.
- *
- * @param {object} [opts] - Optional. When opts.ctx is not provided (old callers), uses legacy behavior (always delete state/artifacts).
- * @param {object} [opts.ctx] - Workflow context (config, etc.)
- * @param {Array} [opts.issues] - Current issue queue for backup pruning
- * @param {boolean} [opts.destructiveReset] - When true, delete state, artifacts, and all backups
- */
-export function ensureCleanLoopStart(
-  workspaceDir,
-  repoRoot,
-  defaultBranch,
-  log,
-  knownBranches = new Set(),
-  opts = {},
-) {
-  const cleaned = {
-    state: false,
-    artifacts: false,
-    branch: false,
-    wipCommitted: false,
-    worktree: false,
-  };
-
-  const ctx = opts.ctx;
-  const destructiveReset = opts.destructiveReset === true;
-  const resumeEnabled =
-    ctx && ctx.config?.workflow?.resumeStepState !== false && !destructiveReset;
-
-  // 1. Delete stale per-issue state and artifacts (or preserve when resume enabled)
-  if (!resumeEnabled) {
-    const sp = statePathFor(workspaceDir);
-    if (existsSync(sp)) cleaned.state = true;
-    const artifactsDir = path.join(workspaceDir, ".coder", "artifacts");
-    for (const name of [
-      "ISSUE.md",
-      "PLAN.md",
-      "PLANREVIEW.md",
-      "REVIEW_FINDINGS.md",
-    ]) {
-      if (existsSync(path.join(artifactsDir, name))) {
-        cleaned.artifacts = true;
-        break;
-      }
-    }
-    clearStateAndArtifacts(workspaceDir);
-  } else {
-    log({ event: "loop_startup_resume_preserved" });
-  }
-
-  // 2. destructiveReset: delete all backups
-  if (destructiveReset) {
-    const backupsDir = path.join(workspaceDir, ".coder", "backups");
-    if (existsSync(backupsDir)) {
-      rmSync(backupsDir, { recursive: true, force: true });
-    }
-  }
-  // 3. Prune orphan backups (issues no longer in queue)
-  else if (resumeEnabled && opts.issues && Array.isArray(opts.issues)) {
-    const validKeys = new Set();
-    const normRepo = (p) => (p ?? ".").trim() || ".";
-    for (const i of opts.issues) {
-      validKeys.add(backupKeyFor(i));
-      if (normRepo(i.repo_path) !== ".") {
-        validKeys.add(backupKeyFor({ ...i, repo_path: "." }));
-      }
-    }
-    const backupsDir = path.join(workspaceDir, ".coder", "backups");
-    if (existsSync(backupsDir)) {
-      try {
-        const entries = readdirSync(backupsDir, { withFileTypes: true });
-        for (const e of entries) {
-          if (e.isDirectory() && !validKeys.has(e.name))
-            rmSync(path.join(backupsDir, e.name), {
-              recursive: true,
-              force: true,
-            });
-        }
-      } catch {
-        // Best-effort prune
-      }
-    }
-  }
-
-  // 3b. Prune orphan checkpoints (runIds no longer in issue queue)
-  if (resumeEnabled && opts.issues && Array.isArray(opts.issues)) {
-    const validRunIds = new Set(
-      opts.issues.map((q) => q?.lastFailedRunId).filter(Boolean),
-    );
-    const coderDir = path.join(workspaceDir, ".coder");
-    if (existsSync(coderDir)) {
-      try {
-        const entries = readdirSync(coderDir, { withFileTypes: true });
-        for (const e of entries) {
-          if (
-            e.isFile() &&
-            e.name.startsWith("checkpoint-") &&
-            e.name.endsWith(".json")
-          ) {
-            const runId = e.name.slice("checkpoint-".length, -".json".length);
-            if (!validRunIds.has(runId))
-              rmSync(path.join(coderDir, e.name), { force: true });
-          }
-        }
-      } catch {
-        // Best-effort prune
-      }
-    }
-  }
-
-  // 4. Ensure git is on the default branch
-  const branchRes = spawnSync("git", ["rev-parse", "--abbrev-ref", "HEAD"], {
-    cwd: repoRoot,
-    encoding: "utf8",
-  });
-  if (branchRes.status !== 0) {
-    const err = (branchRes.stderr || "").trim().slice(0, 200);
-    log({
-      event: "loop_startup_cleanup_failed",
-      step: "detect_branch",
-      error: err,
-    });
-    throw new Error(
-      `Loop startup cleanup failed: could not detect current branch: ${err}`,
-    );
-  }
-  const currentBranch = (branchRes.stdout || "").trim();
-  if (currentBranch && currentBranch !== defaultBranch) {
-    const wipStatus = spawnSync("git", ["status", "--porcelain"], {
-      cwd: repoRoot,
-      encoding: "utf8",
-    });
-    const hasDirty = !!(wipStatus.stdout || "").trim();
-
-    if (hasDirty && knownBranches.has(currentBranch)) {
-      // Agent-managed branch from a prior run: preserve uncommitted WIP
-      const addRes = spawnSync("git", ["add", "-A"], {
-        cwd: repoRoot,
-        encoding: "utf8",
-      });
-      if (addRes.status !== 0) {
-        throw new Error(
-          `Loop startup cleanup failed: git add failed: ${(addRes.stderr || "").trim().slice(0, 200)}`,
-        );
-      }
-      const commitRes = spawnSync(
-        "git",
-        ["commit", "-m", `wip: interrupted work on ${currentBranch}`],
-        { cwd: repoRoot, encoding: "utf8" },
-      );
-      if (commitRes.status === 0) {
-        cleaned.wipCommitted = true;
-      } else {
-        throw new Error(
-          `Loop startup cleanup failed: could not preserve WIP on ${currentBranch} (commit failed): ${(commitRes.stderr || "").trim().slice(0, 150)}`,
-        );
-      }
-    } else if (hasDirty) {
-      const discardOk = discardWorktreeChanges(repoRoot);
-      if (!discardOk) {
-        log({
-          event: "loop_startup_cleanup_failed",
-          step: "discard_unknown_branch",
-          error: "could not discard worktree",
-        });
-        throw new Error(
-          "Loop startup cleanup failed: could not discard worktree on unknown branch",
-        );
-      }
-    }
-
-    const coRes = spawnSync("git", ["checkout", defaultBranch], {
-      cwd: repoRoot,
-      encoding: "utf8",
-    });
-    if (coRes.status !== 0) {
-      const err = (coRes.stderr || "").trim().slice(0, 200);
-      log({
-        event: "loop_startup_cleanup_failed",
-        step: "checkout_default_branch",
-        error: err,
-      });
-      throw new Error(
-        `Loop startup cleanup failed: could not checkout ${defaultBranch}: ${err}`,
-      );
-    }
-    cleaned.branch = true;
-  }
-
-  // 4. Clean any remaining dirty files on the default branch
-  const status = spawnSync("git", ["status", "--porcelain"], {
-    cwd: repoRoot,
-    encoding: "utf8",
-  });
-  const dirtyLines = (status.stdout || "")
-    .split("\n")
-    .filter((l) => l.trim() && !l.slice(3).startsWith(".coder/"));
-  if (dirtyLines.length > 0) {
-    const ok = discardWorktreeChanges(repoRoot);
-    if (!ok) {
-      log({
-        event: "loop_startup_cleanup_failed",
-        step: "clean_worktree",
-        error: "discardWorktreeChanges failed",
-      });
-      throw new Error("Loop startup cleanup failed: could not clean worktree");
-    }
-    cleaned.worktree = true;
-  }
-
-  if (
-    cleaned.state ||
-    cleaned.artifacts ||
-    cleaned.branch ||
-    cleaned.wipCommitted ||
-    cleaned.worktree
-  ) {
-    log({
-      event: "loop_startup_cleanup",
-      ...cleaned,
-      ...(cleaned.branch && { previousBranch: currentBranch }),
-    });
-  }
-}
-
-/**
- * Build args for glab mr list. Exported for testing.
- * Per docs.gitlab.com/cli/mr/list: default is open MRs; --state is not a valid flag.
- * @returns {string[]}
- */
-export function glabMrListArgs() {
-  return ["mr", "list", "--output", "json"];
-}
-
-/**
- * Fetch open PR/MR branches and their diff stats from the hosting platform.
- * Returns an array of { branch, issueId, title, diffStat } suitable for
- * activeBranches. Best-effort: returns [] on failure.
- *
- * @param {string} repoRoot
- * @param {string} defaultBranch
- * @param {(e: object) => void} log
- * @returns {Array<{ branch: string, issueId: string, title: string, diffStat: string }>}
- */
-export function fetchOpenPrBranches(repoRoot, defaultBranch, log) {
-  try {
-    const platform = detectRemoteType(repoRoot);
-    let prs;
-
-    if (platform === "gitlab") {
-      const res = spawnSync("glab", glabMrListArgs(), {
-        cwd: repoRoot,
-        encoding: "utf8",
-        timeout: 15000,
-      });
-      if (res.status !== 0 || !res.stdout) {
-        if (log)
-          log({
-            event: "open_prs_fetch_failed",
-            error: (res.stderr || "glab failed").trim(),
-          });
-        return [];
-      }
-      const mrs = JSON.parse(res.stdout);
-      prs = (Array.isArray(mrs) ? mrs : []).map((mr) => ({
-        branch: mr.source_branch,
-        id: `!${mr.iid}`,
-        title: mr.title || "",
-        fetchRef: `refs/merge-requests/${mr.iid}/head`,
-      }));
-    } else {
-      const res = spawnSync(
-        "gh",
-        [
-          "pr",
-          "list",
-          "--state",
-          "open",
-          "--json",
-          "headRefName,title,number",
-          "--limit",
-          "50",
-        ],
-        { cwd: repoRoot, encoding: "utf8", timeout: 15000 },
-      );
-      if (res.status !== 0 || !res.stdout) {
-        if (log)
-          log({
-            event: "open_prs_fetch_failed",
-            error: (res.stderr || "gh failed").trim(),
-          });
-        return [];
-      }
-      const items = JSON.parse(res.stdout);
-      prs = (Array.isArray(items) ? items : []).map((pr) => ({
-        branch: pr.headRefName,
-        id: `#${pr.number}`,
-        title: pr.title || "",
-        fetchRef: `pull/${pr.number}/head`,
-      }));
-    }
-
-    const result = [];
-    for (const pr of prs) {
-      // Use the platform-specific PR ref to fetch; this works for both
-      // same-repo and fork-based PRs without needing the branch on origin.
-      const localRef = `pr-fetch/${pr.id}`;
-      const fetchRes = spawnSync(
-        "git",
-        ["fetch", "origin", `${pr.fetchRef}:${localRef}`],
-        { cwd: repoRoot, encoding: "utf8", timeout: 10000 },
-      );
-      if (fetchRes.status !== 0) continue;
-      const stat = spawnSync(
-        "git",
-        ["diff", "--stat", `${defaultBranch}...${localRef}`],
-        { cwd: repoRoot, encoding: "utf8", timeout: 10000 },
-      );
-      if (stat.status !== 0) continue;
-      const diffStat = (stat.stdout || "").trim();
-      if (!diffStat) continue;
-      result.push({
-        branch: pr.branch,
-        issueId: pr.id,
-        title: pr.title,
-        diffStat,
-      });
-    }
-
-    if (log) {
-      log({ event: "open_prs_fetched", platform, count: result.length });
-    }
-    return result;
-  } catch (err) {
-    if (log) {
-      log({ event: "open_prs_fetch_failed", error: err.message });
-    }
-    return [];
-  }
-}
+export {
+  ensureCleanLoopStart,
+  fetchOpenPrBranches,
+  glabMrListArgs,
+  resetForNextIssue,
+};
 
 /**
  * Run the autonomous develop loop — process multiple issues.
@@ -2143,122 +1772,6 @@ function hasTrackedFiles(repoRoot) {
     stdio: ["pipe", "pipe", "pipe"],
   });
   return res.status === 0;
-}
-
-/**
- * Reset workspace for next issue in autonomous loop.
- */
-export async function resetForNextIssue(
-  workspaceDir,
-  repoPath,
-  { destructiveReset = false, issueStatus = "completed" } = {},
-) {
-  // Delete per-issue state
-  const statePath = statePathFor(workspaceDir);
-  if (existsSync(statePath)) rmSync(statePath, { force: true });
-
-  // Delete workflow artifacts
-  const artifactsDir = path.join(workspaceDir, ".coder", "artifacts");
-  for (const name of [
-    "ISSUE.md",
-    "PLAN.md",
-    "PLANREVIEW.md",
-    "REVIEW_FINDINGS.md",
-  ]) {
-    const p = path.join(artifactsDir, name);
-    if (existsSync(p)) rmSync(p, { force: true });
-  }
-
-  // Git cleanup
-  const repoRoot = resolveRepoRoot(workspaceDir, repoPath);
-  if (existsSync(repoRoot)) {
-    const preStatus = spawnSync("git", ["status", "--porcelain"], {
-      cwd: repoRoot,
-      encoding: "utf8",
-    });
-    const hasDirtyFiles = !!(preStatus.stdout || "").trim();
-
-    if (
-      hasDirtyFiles &&
-      (issueStatus === "failed" || issueStatus === "skipped")
-    ) {
-      // Preserve partial work on the issue branch for failed/skipped issues.
-      const addRes = spawnSync("git", ["add", "-A"], {
-        cwd: repoRoot,
-        encoding: "utf8",
-      });
-      if (addRes.status !== 0) {
-        throw new Error(
-          `resetForNextIssue: git add failed: ${(addRes.stderr || "").trim().slice(0, 200)}`,
-        );
-      }
-      const commitRes = spawnSync(
-        "git",
-        ["commit", "-m", `wip: partial work (issue ${issueStatus})`],
-        { cwd: repoRoot, encoding: "utf8" },
-      );
-      if (commitRes.status !== 0) {
-        throw new Error(
-          `resetForNextIssue: could not preserve WIP (commit failed): ${(commitRes.stderr || "").trim().slice(0, 150)}`,
-        );
-      }
-    } else if (hasDirtyFiles) {
-      if (!discardWorktreeChanges(repoRoot)) {
-        throw new Error(
-          "resetForNextIssue: could not discard worktree changes",
-        );
-      }
-    }
-
-    const defaultBranch = detectDefaultBranch(repoRoot);
-    const checkoutRes = spawnSync("git", ["checkout", defaultBranch], {
-      cwd: repoRoot,
-      encoding: "utf8",
-    });
-    if (checkoutRes.status !== 0) {
-      throw new Error(
-        `resetForNextIssue: git checkout ${defaultBranch} failed: ${(checkoutRes.stderr || "").trim().slice(0, 200)}`,
-      );
-    }
-
-    // Always remove untracked files after switching to the default branch
-    // to prevent them from leaking into the next issue's workspace.
-    const cleanRes = spawnSync("git", ["clean", "-fd", "--exclude=.coder/"], {
-      cwd: repoRoot,
-      encoding: "utf8",
-    });
-    if (cleanRes.status !== 0) {
-      throw new Error(
-        `resetForNextIssue: git clean failed: ${(cleanRes.stderr || "").trim().slice(0, 200)}`,
-      );
-    }
-
-    // Always clean untracked files after switching branches so the next
-    // issue starts with a pristine working tree.
-    if (destructiveReset) {
-      // Redundant with discardWorktreeChanges + git clean above when hasDirtyFiles,
-      // but ensures staged/worktree match HEAD when we skipped discard (no dirty files).
-      // Skip when repo has no tracked files (e.g. empty initial commit) — restore
-      // would fail with "pathspec '.' did not match any file(s) known to git".
-      const lsRes = spawnSync("git", ["ls-files"], {
-        cwd: repoRoot,
-        encoding: "utf8",
-      });
-      const hasTrackedFiles = !!(lsRes.stdout || "").trim();
-      if (hasTrackedFiles) {
-        const restoreRes = spawnSync(
-          "git",
-          ["restore", "--staged", "--worktree", "."],
-          { cwd: repoRoot, encoding: "utf8" },
-        );
-        if (restoreRes.status !== 0) {
-          throw new Error(
-            `resetForNextIssue: git restore failed: ${(restoreRes.stderr || "").trim().slice(0, 200)}`,
-          );
-        }
-      }
-    }
-  }
 }
 
 /**
